Moved common withdrawal tests into standalone class
This is in preparation for testing cascading withdrawals at various levels
